+++ packages/jwk/src/keyset.ts
@@ -124,13 +124,12 @@ export class Keyset<K extends Key = Key> implements Iterable<K> {
     }
   }
 
-  findSigningKey(search: Omit<KeySearch, 'use'>): [key: Key, alg: string] {
-    const { kid, alg } = search
+  findKey({ kid, alg, use }: KeySearch): [key: Key, alg: string] {
     const matchingKeys: Key[] = []
 
-    for (const key of this.list({ kid, alg, use: 'sig' })) {
+    for (const key of this.list({ kid, alg, use })) {
       // Not a signing key
-      if (!key.canSign) continue
+      if (!key.isPrivate) continue
 
       // Skip negotiation if a specific "alg" was provided
       if (typeof alg === 'string') return [key, alg]
@@ -157,18 +156,20 @@ export class Keyset<K extends Key = Key> implements Iterable<K> {
       }
     }
 
-    throw new TypeError(`No singing key found for ${kid || alg || '<unknown>'}`)
+    throw new TypeError(
+      `No singing key found for ${kid || alg || use || '<unknown>'}`,
+    )
   }
 
   [Symbol.iterator](): IterableIterator<K> {
     return this.keys.values()
   }
 
-  async sign(
-    { alg: searchAlg, kid: searchKid, ...header }: JwtSignHeader,
+  async createJwt(
+    { alg: sAlg, kid: sKid, ...header }: JwtSignHeader,
     payload: JwtPayload | JwtPayloadGetter,
   ) {
-    const [key, alg] = this.findSigningKey({ alg: searchAlg, kid: searchKid })
+    const [key, alg] = this.findKey({ alg: sAlg, kid: sKid, use: 'sig' })
     const protectedHeader = { ...header, alg, kid: key.kid }
 
     if (typeof payload === 'function') {
@@ -178,7 +179,7 @@ export class Keyset<K extends Key = Key> implements Iterable<K> {
     return key.createJwt(protectedHeader, payload)
   }
 
-  async verify<
+  async verifyJwt<
     P extends Record<string, unknown> = JwtPayload,
     C extends string = string,
   >(token: Jwt, options?: VerifyOptions<C>) {
@@ -187,9 +188,10 @@ export class Keyset<K extends Key = Key> implements Iterable<K> {
 
     const errors: unknown[] = []
 
-    for (const key of this.list({ use: 'sig', kid, alg })) {
+    for (const key of this.list({ kid, alg })) {
       try {
-        return await key.verifyJwt<P, C>(token, options)
+        const result = await key.verifyJwt<P, C>(token, options)
+        return { ...result, key }
       } catch (err) {
         errors.push(err)
       }