+++ packages/oauth-client-browser-example/vite.config.ts
@@ -10,17 +10,13 @@ export default defineConfig({
   },
   optimizeDeps: {
     include: [
+      '@atproto/api',
       '@atproto/oauth-client',
       '@atproto/oauth-client-browser',
       '@atproto/oauth-client-metadata',
       '@atproto/oauth-server-metadata',
+      '@atproto/xrpc',
     ],
   },
-  plugins: [
-    react({
-      babel: {
-        plugins: ['@babel/plugin-syntax-import-assertions'],
-      },
-    }),
-  ],
+  plugins: [react()],
 })